# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Planet Born API — واجهة «كوكب يولد أمامك»
  version: 0.1.0
  description: |
    REST API for the living planet platform. Auth via Bearer JWT (short-lived)
    + rotated refresh tokens. Realtime deltas over Socket.IO (see docs/WEBSOCKETS.md).
servers:
  - url: http://localhost:4000
...
paths:
...
  /auth/register:
    post:
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterInput" }
      responses:
        "200": { description: AuthResponse, content: { application/json: { schema: { $ref: "#/components/schemas/AuthResponse" } } } }
        "409": { description: Email already registered }
        "429": { description: Rate limited }
  /auth/login:
    post:
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200": { description: AuthResponse, content: { application/json: { schema: { $ref: "#/components/schemas/AuthResponse" } } } }
        "401": { description: Invalid credentials }
  /auth/refresh:
    post:
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200": { description: Rotated token pair, content: { application/json: { schema: { $ref: "#/components/schemas/AuthResponse" } } } }
        "401": { description: Invalid / expired / reuse-detected }
...
  /auth/oauth:
    post:
      tags: [auth]
      security: []
      description: Google/Apple sign-in. 400 `oauth_not_configured` when keys are absent (documented sandbox behavior).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider, idToken]
              properties:
                provider: { type: string, enum: [google, apple] }
                idToken: { type: string }
      responses:
        "200": { description: AuthResponse }
        "400": { description: oauth_not_configured }
...
  /planets/current:
    get:
      tags: [planets]
      security: []
      responses:
        "200": { description: PlanetSummary, content: { application/json: { schema: { $ref: "#/components/schemas/PlanetSummary" } } } }
        "404": { description: no_planet (run db:seed) }
  /planets/{id}/regions:
    get:
      tags: [planets]
      security: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: biome, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 60 } }
      responses:
        "200": { description: Regions list }
  /planets/{id}/regions/{cellId}:
    get:
      tags: [planets]
      security: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: cellId, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: RegionInfo }
  /planets/{id}/maps/{layer}:
    get:
      tags: [planets]
      security: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: layer, in: path, required: true, schema: { type: string, enum: [height, biome, temperature, moisture, pollution, political, vegetation, nightlights] } }
        - { name: tick, in: query, schema: { type: integer }, description: Latest layer at or before this tick (era viewing) }
      responses:
        "200": { description: "{ layer, tick, pngBase64 }" }
...
  /planets/{id}/species:
    get:
      tags: [planets]
      security: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
      responses:
        "200": { description: Paginated species }
...
  /planets/{id}/snapshots/compare:
    get:
      tags: [planets]
      security: []
      parameters:
        - { name: from, in: query, required: true, schema: { type: integer } }
        - { name: to, in: query, required: true, schema: { type: integer } }
      responses:
        "200": { description: Metric deltas between two snapshots }
  /planets/{id}/events:
    get:
      tags: [planets]
      security: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 20 } }
        - { name: types, in: query, schema: { type: string }, description: CSV of event types }
        - { name: minImportance, in: query, schema: { type: integer, minimum: 1, maximum: 5 } }
        - { name: civId, in: query, schema: { type: string } }
        - { name: contributionId, in: query, schema: { type: string } }
        - { name: fromTick, in: query, schema: { type: integer } }
        - { name: toTick, in: query, schema: { type: integer } }
      responses:
        "200": { description: Paginated world events }
...
  /contributions/drafts:
    post:
      tags: [contributions]
      description: Step 1-2 of the wizard. Rate limited 20/hour.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ContributionDraftInput" }
      responses:
        "200": { description: Draft contribution }
        "422": { description: moderation_blocked }
...
  /contributions/{id}/analyze:
    post:
      tags: [contributions]
      description: AI parse -> structured traits -> balance -> causal prior.
      responses:
        "200": { description: "{ contribution, balance, graph, provider, sandbox }" }
        "422": { description: moderation_blocked }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterInput:
      type: object
      required: [email, password, displayName]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 10 }
        displayName: { type: string, minLength: 2, maxLength: 60 }
        locale: { type: string, enum: [ar, en] }
    AuthResponse:
      type: object
      properties:
        user: { $ref: "#/components/schemas/User" }
        tokens:
          type: object
          properties:
            accessToken: { type: string }
            refreshToken: { type: string }
            expiresIn: { type: integer }
            tokenType: { type: string, example: Bearer }
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        displayName: { type: string }
        role: { type: string, enum: [visitor, user, explorer, life_maker, civ_builder, historian, moderator, sim_manager, admin, super_admin] }
        locale: { type: string }
    PlanetSummary:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        seed: { type: integer }
        tick: { type: integer }
        status: { type: string, enum: [running, paused] }
        civilizations: { type: integer }
        cities: { type: integer }
        species: { type: integer }
        plants: { type: integer }
        technologies: { type: integer }
        activeWars: { type: integer }
        eventCount: { type: integer }
        globalTempShift: { type: number }
    ContributionDraftInput:
      type: object
      required: [planetId, text]
      properties:
        planetId: { type: string }
        text: { type: string, minLength: 2, maxLength: 1200 }
        category: { type: string, enum: [creature, plant, resource, civilization, invention, natural_phenomenon, disease, culture, world_law, custom] }
security:
  - bearerAuth: []